+++ sql/GeneraliDB/Tables/dbo.IssReports.sql
@@ -1,0 +1,21 @@
+﻿USE [Generali]
+GO
+DROP TABLE [dbo].[IssReports]
+GO
+SET ANSI_NULLS ON
+GO
+SET QUOTED_IDENTIFIER ON
+GO
+CREATE TABLE [dbo].[IssReports](
+	[ID] [int] IDENTITY(1,1) NOT NULL,
+	[ReportForDate] [date] NULL,
+	[ReportTimeStamp] [datetime] NULL,
+	[ReportByUserID] [int] NULL,
+	[OnTime] [bit] NULL,
+	[Category] [nvarchar](50) NULL,
+ CONSTRAINT [PK_IssReports] PRIMARY KEY CLUSTERED 
+(
+	[ID] ASC
+)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
+) ON [PRIMARY]
+GO